backend-py/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.31
# Using sync SQLite - no additional driver needed
python-dotenv==1.0.1
ldap3==2.9.1
PyJWT==2.8.0
cryptography==42.0.8
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
email-validator==2.2.0
paramiko==3.4.0 